Add javadocs and a little cleanup
