+++ README.md
@@ -1,46 +1,48 @@
-## Hashgraph application template
+# Hashgraph application template
 This is an alternative (arguably friendlier) maven setup for Hashgraph SDK based applications.
 
 It allows you to build and run hashgraph apps using the SDK more like an external resource rather than creating new modules along the demo apps inside the SDK directory structure.
 
 Clone this repo to whatever location (independend from the Hashgraph sdk).
-#### Install Hashgraph SDK
-Follow the pre-requisites, download and unzip the SDK to preferred location. See https://dev.hashgraph.com/docs/installation/
+## Install Hashgraph SDK
+Follow the pre-requisites, download and unzip the SDK to preferred location. See https://www.swirlds.com/download/
 
-#### Configure the SDK:
+### Configure the SDK:
 1. change the `config.txt` file found in the SDK as follows:
    * comment the GameDemo.jar line and add a new app line for this application (HelloSwirld.jar)
 ```
 ...
 # app,        GameDemo.jar
-app,        HelloSwirld.jar
+app,          HelloSwirld.jar
 ...
 ```
    * Switch off the TLS encryption if you find that the startup is slow. Just uncomment the line:
 ```
 TLS, off
 ```
-#### Configure the pom.xml of this project:
+
+## Configure this Project
 You only need to change the sdk.dir property to refer to the path of your sdk installation (the directory path where you found the config.txt file).
 The default is:
 ```
-<sdk.dir>/home/swirlds/sdk</sdk.dir>
+<swirlds.sdk>/home/swirlds/sdk</swirlds.sdk>
 ```
-#### Build it
+
+### Building
 ```
 mvn clean install
 ```
 this will package the app jar and copy it to the `apps` dir inside the hashgraph sdk
 
-#### Run from IntelliJ IDEA
+### Run from IntelliJ IDEA
 1. Run -> Edit Configurations...
 2. Add new Application cofiguration
-3. Main Class: HelloSwirldMain
+3. Main Class: Main
 4. Working Directory: the sdk path (the same one you configured in the pom.xml)
 
 You should see four console windows and one main browser window.
 
-#### Run from command line
+### Run from command line
 To run it in the usual way, go to hashgraph sdk dir
 ```
 java -jar swirlds.jar
